+++ src/geometry/shape.rs
@@ -2,7 +2,7 @@ use crate::prelude::*;
 
 pub trait Shape {
     fn normal_at(&self, world_space_point: Point) -> Vector;
-    fn times_of_intersections_with<'s, 'r>(&'s self, ray: &'r Ray) -> Vec<f32>
+    fn times_of_intersections_with<'s, 'r>(&'s self, ray: &'r Ray) -> Vec<f64>
     where
         'r: 's;
     fn transform(&self) -> &Transform;
